Move helper tests into new utility namespace
